test: 月次予算テストの現在年月生成を共通化
3つのテストに重複していた現在年月とeffective dateの生成を既存のbudget test utilityへ集約する。
